+++ src/main/test/algothief/PoliciaTest.java
@@ -1,5 +1,5 @@
-package edu.fiuba.algo3.modelo;
-
+package algothief;
+import edu.fiuba.algo3.modelo.*;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
