+++ src/main/test/algothief/UbicacionTest.java
@@ -1,0 +1,5 @@
+package algothief;
+
+public class UbicacionTest {
+    
+}
